requirements.txt
matplotlib>=3.8
seaborn>=0.13
numpy>=1.26
pyyaml>=6.0
pytest>=7.4
